Merge branch 'master' of
https://github.com/skyvault05/houseoftomorrow.git

Conflicts:
	src/main/java/hot/member/domain/Bathroom.java
	src/main/java/hot/member/domain/Constructor.java
	src/main/java/hot/member/domain/Estimate.java
	src/main/java/hot/member/domain/Floor.java
	src/main/java/hot/member/domain/Kitchen.java
	src/main/java/hot/member/domain/Linoleum.java
	src/main/java/hot/member/domain/Member.java
	src/main/java/hot/member/domain/Papering.java
	src/main/java/hot/member/domain/Tile.java
	src/main/java/hot/member/repository/ConstructorRepository.java
	src/main/java/hot/member/repository/EstimateRepository.java
	src/test/java/hot/HouseoftomorrowApplicationTests.java

+++ src/main/java/hot/member/domain/CommCategory.java
@@ -26,7 +26,7 @@ public class CommCategory {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "comm_category_no")
-	private Long commCategoryNo;
+	private Integer commCategoryNo;
 	@Column(name = "comm_category_name")
 	private String commCategoryName;
