+++ src/main/java/hot/member/repository/CommCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import hot.member.domain.CommCategory;
 
-public interface CommCategoryRepository extends JpaRepository<CommCategory, Long> {
+public interface CommCategoryRepository extends JpaRepository<CommCategory, Integer> {
 
 }
